Move Booking utils
As with the previous commit where we moved the premises utils to their
own folder we now do the same with booking utils
