+++ server/utils/bookings/index.ts
@@ -9,15 +9,15 @@ import type {
 } from '@approved-premises/ui'
 import type { BedSummary, Booking, BookingStatus, BookingSummary, PremisesBooking } from '@approved-premises/api'
 import { addDays, isBefore, isEqual, isWithinInterval } from 'date-fns'
-import paths from '../paths/manage'
-import applyPaths from '../paths/apply'
-import assessPaths from '../paths/assess'
-import { DateFormats, todayAtMidnight } from './dateUtils'
-import { SanitisedError } from '../sanitisedError'
-import { linebreaksToParagraphs, linkTo } from './utils'
-import { isFullPerson, laoName } from './personUtils'
-import { convertObjectsToRadioItems } from './formUtils'
-import { StatusTag, StatusTagOptions } from './statusTag'
+import paths from '../../paths/manage'
+import applyPaths from '../../paths/apply'
+import assessPaths from '../../paths/assess'
+import { DateFormats, todayAtMidnight } from '../dateUtils'
+import { SanitisedError } from '../../sanitisedError'
+import { linebreaksToParagraphs, linkTo } from '../utils'
+import { isFullPerson, laoName } from '../personUtils'
+import { convertObjectsToRadioItems } from '../formUtils'
+import { StatusTag, StatusTagOptions } from '../statusTag'
 
 const UPCOMING_WINDOW_IN_DAYS = 365 * 10
