+++ server/utils/nunjucksSetup.ts
@@ -45,7 +45,7 @@ import * as OASysUtils from './assessments/oasysUtils'
 import * as OffenceUtils from './offenceUtils'
 import * as AttachDocumentsUtils from './attachDocumentsUtils'
 import * as OasysImportUtils from './oasysImportUtils'
-import * as BookingUtils from './bookingUtils'
+import * as BookingUtils from './bookings'
 import * as TasklistUtils from './taskListUtils'
 import * as FormUtils from './formUtils'
 import * as UserUtils from './users'
